Broke down GetHeatmapMonth into GetHeatmapMonth + GetHeatMapWeek
